+++ game_patch/input/key.cpp
@@ -123,6 +123,24 @@ CodeInjection key_get_hook{
     },
 };
 
+CodeInjection key_down_handler_injection{
+    0x0051E9AC,
+    [] (auto& regs) {
+        const int virtual_key = addr_as_ref<int>(regs.esp + 4);
+        // For numeric keypads, we need to fix these keys' scan codes.
+        auto& scan_code = regs.eax;
+        if (virtual_key == VK_PRIOR) {
+            scan_code = rf::KEY_PAGEUP;
+        } else if (virtual_key == VK_NEXT) {
+            scan_code = rf::KEY_PAGEDOWN;
+        } else if (virtual_key == VK_END) {
+            scan_code = rf::KEY_END;
+        } else if (virtual_key == VK_HOME) {
+            scan_code = rf::KEY_HOME;
+        }
+    },
+};
+
 void key_apply_patch()
 {
     // Support non-US keyboard layouts
@@ -135,4 +153,6 @@ void key_apply_patch()
 
     // win32 console support and addition of os_poll
     key_get_hook.install();
+
+    key_down_handler_injection.install();
 }